requirements.txt
fastapi
uvicorn[standard]
requests
feedparser
google-generativeai
firebase-admin